[WIP] Refactor code in VPC to reduce code redundancy
