+++ aws/resources/ec2_internet_gateway.go
@@ -2,9 +2,12 @@ package resources
 
 import (
 	"context"
+	"fmt"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
+	awsgo "github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/gruntwork-io/cloud-nuke/config"
 	"github.com/gruntwork-io/cloud-nuke/logging"
@@ -18,8 +21,8 @@ func (igw *InternetGateway) setFirstSeenTag(gateway ec2.InternetGateway, value t
 		Resources: []*string{gateway.InternetGatewayId},
 		Tags: []*ec2.Tag{
 			{
-				Key:   aws.String(util.FirstSeenTagKey),
-				Value: aws.String(util.FormatTimestamp(value)),
+				Key:   awsgo.String(util.FirstSeenTagKey),
+				Value: awsgo.String(util.FormatTimestamp(value)),
 			},
 		},
 	})
@@ -109,60 +112,58 @@ func (igw *InternetGateway) getAll(_ context.Context, configObj config.Config) (
 	return identifiers, nil
 }
 
-func (igw *InternetGateway) nuke(id *string) error {
-
-	// detaching the gateway from attached vpcs
-	if err := igw.detachInternetGateway(id); err != nil {
-		return errors.WithStackTrace(err)
+func nukeInternetGateway(client ec2iface.EC2API, vpcID string) error {
+	logging.Debug(fmt.Sprintf("Start nuking Internet Gateway for vpc: %s", vpcID))
+	input := &ec2.DescribeInternetGatewaysInput{
+		Filters: []*ec2.Filter{
+			{
+				Name:   awsgo.String("attachment.vpc-id"),
+				Values: []*string{awsgo.String(vpcID)},
+			},
+		},
 	}
-
-	// nuking the gateway
-	if err := igw.nukeInternetGateway(id); err != nil {
-		return errors.WithStackTrace(err)
-	}
-
-	return nil
-}
-
-func (igw *InternetGateway) nukeInternetGateway(id *string) error {
-	logging.Debugf("Deleting Internet gateway %s", *id)
-
-	_, err := igw.Client.DeleteInternetGateway(&ec2.DeleteInternetGatewayInput{
-		InternetGatewayId: id,
-	})
+	igw, err := client.DescribeInternetGateways(input)
 	if err != nil {
-		logging.Debugf("[Failed] Error deleting internet gateway %s: %s", *id, err)
+		logging.Debug(fmt.Sprintf("Failed to describe internet gateways for vpc: %s", vpcID))
 		return errors.WithStackTrace(err)
 	}
-	logging.Debugf("[Ok] internet gateway deleted successfully %s", *id)
 
-	return nil
-}
+	if len(igw.InternetGateways) < 1 {
+		logging.Debug(fmt.Sprintf("No Internet Gateway to delete."))
+		return nil
+	}
 
-func (igw *InternetGateway) detachInternetGateway(id *string) error {
-	logging.Debugf("Detaching Internet gateway %s", *id)
-
-	output, err := igw.Client.DescribeInternetGateways(&ec2.DescribeInternetGatewaysInput{
-		InternetGatewayIds: []*string{id},
-	})
+	logging.Debug(fmt.Sprintf("Detaching Internet Gateway %s",
+		awsgo.StringValue(igw.InternetGateways[0].InternetGatewayId)))
+	_, err = client.DetachInternetGateway(
+		&ec2.DetachInternetGatewayInput{
+			InternetGatewayId: igw.InternetGateways[0].InternetGatewayId,
+			VpcId:             awsgo.String(vpcID),
+		},
+	)
 	if err != nil {
-		logging.Debugf("[Failed] Error describing internet gateway %s: %s", *id, err)
+		logging.Debug(fmt.Sprintf("Failed to detach internet gateway %s",
+			awsgo.StringValue(igw.InternetGateways[0].InternetGatewayId)))
 		return errors.WithStackTrace(err)
 	}
+	logging.Debug(fmt.Sprintf("Successfully detached internet gateway %s",
+		awsgo.StringValue(igw.InternetGateways[0].InternetGatewayId)))
 
-	for _, gateway := range output.InternetGateways {
-		for _, attachments := range gateway.Attachments {
-			_, err := igw.Client.DetachInternetGateway(&ec2.DetachInternetGatewayInput{
-				InternetGatewayId: id,
-				VpcId:             attachments.VpcId,
-			})
-			if err != nil {
-				logging.Debugf("[Failed] Error detaching internet gateway %s: %s", *id, err)
-				return errors.WithStackTrace(err)
-			}
-		}
+	logging.Debug(fmt.Sprintf("Deleting internet gateway %s",
+		awsgo.StringValue(igw.InternetGateways[0].InternetGatewayId)))
+	_, err = client.DeleteInternetGateway(
+		&ec2.DeleteInternetGatewayInput{
+			InternetGatewayId: igw.InternetGateways[0].InternetGatewayId,
+		},
+	)
+	if err != nil {
+		logging.Debug(fmt.Sprintf("Failed to delete internet gateway %s",
+			awsgo.StringValue(igw.InternetGateways[0].InternetGatewayId)))
+		return errors.WithStackTrace(err)
 	}
-	logging.Debugf("[Ok] internet gateway detached successfully %s", *id)
+	logging.Debug(fmt.Sprintf("Successfully deleted internet gateway %s",
+		awsgo.StringValue(igw.InternetGateways[0].InternetGatewayId)))
+
 	return nil
 }
 
@@ -181,10 +182,10 @@ func (igw *InternetGateway) nukeAll(identifiers []*string) error {
 			continue
 		}
 
-		err := igw.nuke(id)
+		err := nuke(igw.Client, *id)
 		// Record status of this resource
 		e := r.Entry{ // Use the 'r' alias to refer to the package
-			Identifier:   aws.StringValue(id),
+			Identifier:   awsgo.StringValue(id),
 			ResourceType: "Internet Gateway",
 			Error:        err,
 		}
